rename Counter to CounterControls
